pyproject.toml
[project]
name = "multimodal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-google-genai>=2.1.8",
    "langchain-ollama>=0.3.6",
    "pdf2image>=1.17.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "pytesseract>=0.3.13",
    "python-magic>=0.4.27",
    "unstructured[pdf]>=0.18.11",
    "uvicorn>=0.35.0",
]
